test:嘗試修復校園活動讀取
